cleanup top-level test files
