--- packages/slate-react/test/utils/index.js
@@ -1,7 +1,0 @@
-import { KeyUtils } from 'slate'
-
-beforeEach(KeyUtils.resetGenerator)
-
-describe('utils', () => {
-  // require('./get-children-decorations')
-})
